update boss tutorial

+++ example/del_jobs.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
-
 function delete_job() {
   jobname=$1
+  if [[ "$jobname" == "" ]]; then
+      echo "Usage: sh edl_jobs.sh [all|<job-name>]"
+      exit 0
+  fi
   kubectl delete trainingjob $jobname
   kubectl delete job $jobname-trainer
   kubectl delete rs $jobname-master $jobname-pserver
@@ -9,7 +12,7 @@ function delete_job() {
 
 function delete_all() {
   jobs=$(kubectl get trainingjob | tail -n +2 | awk '{print $1}')
-  for job in "${jobs[@]}"
+  for job in ${jobs[@]}
   do
     delete_job $job
   done
